FIX logic error in Val::is()
certain checks would not be correctly performed if a class
by the same name as the given check existed.

For instance, if the class »Throwable« exists, then the "throwable"
check would not be run because is_a($foo, 'throwable') would be true
due to the fact that PHP class names are (partially) case-insensitive.